+++ extension/parquet/include/reader/variant_column_reader.hpp
@@ -36,8 +36,6 @@ public:
 	uint64_t TotalCompressedSize() override;
 	void RegisterPrefetch(ThriftFileTransport &transport, bool allow_merge) override;
 	static bool TypedValueLayoutToType(const LogicalType &typed_value, LogicalType &logical_type);
-	//! Inverse of GetTransformFunction: decode a binary Variant value (metadata followed by value) into a VARIANT.
-	static ScalarFunction GetBytesToVariantFunction();
 
 protected:
 	idx_t metadata_reader_idx;
